+++ docs/changelog.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased](https://github.com/JoshKarpel/waxy/compare/v0.4.0...HEAD)
+## [Unreleased](https://github.com/JoshKarpel/waxy/compare/v0.5.0...HEAD)
+
+## [0.5.0](https://github.com/JoshKarpel/waxy/compare/v0.4.0...v0.5.0) - 2026-04-03
+
+### Added
+
+- [#45](https://github.com/JoshKarpel/waxy/pull/45) `Rect.intersection(other)` returns the intersection rectangle (including edge-touching degenerate intersections), or `None` if the two rectangles do not intersect.
 
 ## [0.4.0](https://github.com/JoshKarpel/waxy/compare/v0.3.0...v0.4.0) - 2026-03-01